examples/rogueAP/requirements.txt
pika
pyyaml
